Find Extremes in Prices
Used `filter` to filter out invalid prices and converts prices to numbers. `reduce` used to track the minimum and maximum prices in the object. `find` was used to get the product names for the min and max prices. Return for the formatted string and wrapping in an Immediately Invoked Function Expression.